requirements.txt
googlemaps==4.10.0
folium==0.14.0
geopy==2.4.0
pandas==2.0.3
python-dotenv==1.0.0

